requirements.txt
# pip install -r requirements.txt

numpy
pandas
matplotlib
yfinance
openpyxl
python-binance
schedule